requirements.txt
aiocontextvars==0.2.2
aliyun-python-sdk-core==2.13.24
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.10.1
allure-pytest==2.8.18
allure-python-commons==2.8.18
APScheduler==3.6.3
attrs==20.1.0
BeautifulReport==0.1.3
celery==4.4.1
cycler==0.10.0
DBUtils==1.3
ddt==1.4.1
demjson==2.2.4
dictor==0.1.7
diff-match-patch==20200713
Django==3.1.2
django-celery-beat==2.1.0
django-celery-results==1.2.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.1
django-formtools==2.2
django-import-export==0.4.2
django-redis==4.12.1
django-rest-swagger==2.2.0
django-shortuuidfield==0.1.3
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
doctor==3.13.7
drf-extensions==0.6.0
drf-yasg==1.17.1
jmespath==0.10.0
jsonpath==0.82
jsonschema==2.6.0
loguru==0.5.1
openapi-codec==1.3.2
openpyxl==3.0.5
oss2==2.12.1
parameterized==0.7.4
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
PyJWT==1.7.1
pymongo==3.11.0
PyMySQL==0.9.3
pytest==6.0.1
python-crontab==2.5.1
PyYAML==3.13
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
selenium==3.141.0
simplejson==3.17.2
urllib3==1.25.9
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
Faker~=8.8.2